requirements.txt
pytest
numpy
langchain
openai
pydantic
SQLAlchemy
voyageai
python-dotenv 
python-lorem
requests